feat(admin): update AI admin actions and routing

+++ ink-ms-admin/src/main/java/com/inklusport/admin/service/AdminRoleService.java
@@ -150,7 +150,7 @@ public class AdminRoleService {
                 .description(role.getDescription())
                 .createdAt(role.getCreatedAt())
                 .permissions(permissions)
-                .adminCount(adminCount)
+                .adminCount(adminCount.intValue())
                 .build();
     }
 }
